+++ agent/runtime/src/main/java/com/walmartlabs/concord/agent/ExecutionManager.java
@@ -15,8 +15,12 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
 import java.util.*;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
@@ -82,6 +86,9 @@ public class ExecutionManager {
         Map<String, Object> agentParams = getAgentParameters(tmpDir);
         Collection<String> jvmArgs = (Collection<String>) agentParams.getOrDefault(Constants.JVM_ARGS_KEY, DEFAULT_JVM_ARGS);
         log.info("start ['{}'] -> JVM args: {}", instanceId, jvmArgs);
+
+        Collection<String> deps = getDependencies(tmpDir);
+        collectDependencies(instanceId, tmpDir, deps);
 
         synchronized (mutex) {
             statuses.put(instanceId, JobStatus.RUNNING);
@@ -233,5 +240,70 @@ public class ExecutionManager {
         } catch (IOException e) {
             throw new ExecutionException("Error while reading an agent parameters file", e);
         }
+    }
+
+    private static Collection<String> getDependencies(Path payload) throws ExecutionException {
+        Path p = payload.resolve(Constants.REQUEST_DATA_FILE_NAME);
+        if (!Files.exists(p)) {
+            return Collections.emptyList();
+        }
+
+        try (InputStream in = Files.newInputStream(p)) {
+            ObjectMapper om = new ObjectMapper();
+            Map<String, Object> m = om.readValue(in, Map.class);
+
+            Collection<String> deps = (Collection<String>) m.get(Constants.DEPENDENCIES_KEY);
+            return deps != null ? deps : Collections.emptyList();
+        } catch (IOException e) {
+            throw new ExecutionException("Error while reading a list of dependencies", e);
+        }
+    }
+
+    private static void collectDependencies(String instanceId, Path baseDir, Collection<String> deps) throws ExecutionException {
+        if (deps.isEmpty()) {
+            return;
+        }
+
+        Path libDir = baseDir.resolve(Constants.LIBRARIES_DIR_NAME);
+        if (!Files.exists(libDir)) {
+            try {
+                Files.createDirectories(libDir);
+            } catch (IOException e) {
+                throw new ExecutionException("Dependencies processing error", e);
+            }
+        }
+
+        // TODO collect all errors before throwing an exception
+        for (String d : deps) {
+            URL url;
+            try {
+                url = new URL(d);
+            } catch (MalformedURLException e) {
+                throw new ExecutionException("Invalid dependency URL: " + d, e);
+            }
+
+            String name = getLastPart(url);
+            Path dst = libDir.resolve(name);
+            if (Files.exists(dst)) {
+                log.warn("collectDependencies ['{}'] -> library already exists: {}", instanceId, dst);
+                continue;
+            }
+
+            try (InputStream in = url.openStream();
+                 OutputStream out = Files.newOutputStream(dst, StandardOpenOption.CREATE)) {
+                IOUtils.copy(in, out);
+            } catch (IOException e) {
+                throw new ExecutionException("Error while copying a dependency: " + d, e);
+            }
+        }
+    }
+
+    private static String getLastPart(URL url) {
+        String p = url.getPath();
+        int idx = p.lastIndexOf('/');
+        if (idx >= 0 && idx + 1 < p.length()) {
+            return p.substring(idx + 1);
+        }
+        throw new IllegalArgumentException("Invalid URL: " + url);
     }
 }