+++ it/server/src/test/java/com/walmartlabs/concord/it/server/DependenciesIT.java
@@ -1,0 +1,56 @@
+package com.walmartlabs.concord.it.server;
+
+import com.walmartlabs.concord.common.Constants;
+import com.walmartlabs.concord.common.IOUtils;
+import com.walmartlabs.concord.server.api.process.ProcessResource;
+import com.walmartlabs.concord.server.api.process.ProcessStatus;
+import com.walmartlabs.concord.server.api.process.ProcessStatusResponse;
+import com.walmartlabs.concord.server.api.process.StartProcessResponse;
+import org.junit.Test;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Arrays;
+import java.util.zip.ZipOutputStream;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+public class DependenciesIT extends AbstractServerIT {
+
+    @Test(timeout = 30000)
+    public void testUploadAndRun() throws Exception {
+        String dep = "file:///" + ITConstants.DEPENDENCIES_DIR + "/example.jar";
+
+        String request = "{ \"entryPoint\": \"main\", \"dependencies\": [ \"" + dep + "\" ] }";
+        Path tmpDir = Files.createTempDirectory("test");
+        Path requestFile = tmpDir.resolve(Constants.REQUEST_DATA_FILE_NAME);
+        Files.write(requestFile, Arrays.asList(request));
+
+        Path src = Paths.get(DependenciesIT.class.getResource("deps").toURI());
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        try (ZipOutputStream zip = new ZipOutputStream(baos)) {
+            IOUtils.zip(zip, src);
+            IOUtils.zip(zip, tmpDir);
+        }
+
+        byte[] payload = baos.toByteArray();
+
+        // ---
+
+        ProcessResource processResource = proxy(ProcessResource.class);
+        StartProcessResponse spr = processResource.start(new ByteArrayInputStream(payload));
+        assertNotNull(spr.getInstanceId());
+
+        ProcessStatusResponse psr = waitForCompletion(processResource, spr.getInstanceId());
+        assertEquals(ProcessStatus.FINISHED, psr.getStatus());
+
+        // ---
+
+        byte[] ab = getLog(psr.getLogFileName());
+        assertLog(".*Hello!.*", ab);
+    }
+}