+++ plugins/templates/ansible/src/main/filtered-resources/_main.js
@@ -1,5 +1,8 @@
 ({
     entryPoint: "main",
+    dependencies: [
+        "concord:///ansible-tasks-${project.version}.jar"
+    ],
     arguments: {
         ansibleParams: _input
     }