+++ server/impl/src/main/java/com/walmartlabs/concord/server/cfg/DependencyStoreConfiguration.java
@@ -1,16 +1,13 @@
 package com.walmartlabs.concord.server.cfg;
 
-import com.google.common.base.Throwables;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
-import java.io.IOException;
 import java.io.Serializable;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Properties;
 
 @Named
 @Singleton
@@ -20,29 +17,15 @@ public class DependencyStoreConfiguration implements Serializable {
 
     public static final String DEPS_STORE_DIR_KEY = "DEPS_STORE_DIR";
 
-    private static final Path DEFAULT_DIRECTORY;
-
-    static {
-        Properties props = new Properties();
-        try {
-            props.load(DependencyStoreConfiguration.class.getResourceAsStream("maven.properties"));
-        } catch (IOException e) {
-            throw Throwables.propagate(e);
-        }
-
-        String s = props.getProperty("depsDir");
-        if (s == null || s.trim().isEmpty()) {
-            s = System.getProperty("user.dir");
-        }
-
-        DEFAULT_DIRECTORY = Paths.get(s);
-    }
-
     private final Path depsDir;
 
     public DependencyStoreConfiguration() {
         String s = System.getenv(DEPS_STORE_DIR_KEY);
-        this.depsDir = s != null ? Paths.get(s).toAbsolutePath() : DEFAULT_DIRECTORY;
+        if (s == null) {
+            log.warn("init -> {} must be set in order to use project templates and custom process dependencies", DEPS_STORE_DIR_KEY);
+        }
+
+        this.depsDir = s != null ? Paths.get(s).toAbsolutePath() : null;
         log.info("init -> depsDir: {}", depsDir);
     }