+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/SelfContainedArchivePipeline.java
@@ -13,9 +13,9 @@ public class SelfContainedArchivePipeline extends Chain {
     @Inject
     public SelfContainedArchivePipeline(Injector injector) {
         super(injector.getInstance(WorkspaceArchiveProcessor.class),
+                injector.getInstance(RequestDefaultsParsingProcessor.class),
                 injector.getInstance(TemplateProcessor.class),
                 injector.getInstance(DependenciesProcessor.class),
-                injector.getInstance(RequestDefaultsParstingProcessor.class),
                 injector.getInstance(RequestDataStoringProcessor.class),
                 injector.getInstance(RunnerProcessor.class),
                 injector.getInstance(LogFileProcessor.class),