+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/pipelines/processors/RequestDefaultsParsingProcessor.java
@@ -17,7 +17,7 @@ import java.util.Map;
 
 @Named
 @Singleton
-public class RequestDefaultsParstingProcessor implements PayloadProcessor {
+public class RequestDefaultsParsingProcessor implements PayloadProcessor {
 
     private final ObjectMapper objectMapper = new ObjectMapper();